requirements/launchpad.txt
# Please keep requirements in this file sorted alphabetically and
# case-insensitively.  Local forks should use +lpN suffixes and be annotated
# with information about where to find the code for the fork.
#
# Don't list entries from ztk-versions.cfg here unless overriding their
# versions; they will be included automatically.

# lp:~launchpad/ampoule:launchpad
ampoule==19.12.0.post1
amqp==2.6.1
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==0.23.0
attrs==19.3.0
Automat==20.2.0
backports.functools-lru-cache==1.5
# ztk-versions.cfg uses 3.2.0 on Python 3, but that drops support for Python
# 3.5.  Pin to 3.1.7 until we no longer care about xenial.
bcrypt==3.1.7
beautifulsoup4==4.7.1
billiard==3.6.4.0
bleach==3.3.1
bleach-allowlist==1.0.3
breezy==3.2.0
bson==0.5.9
boto3==1.16.63
botocore==1.19.63
celery==4.4.7
Chameleon==3.6.2
configobj==5.0.6
contextvars==2.4
constantly==15.1.0
cookies==2.2.1
cryptography==2.7
cssselect==0.9.1
cssutils==1.0.2
Cython==0.29.17
d2to1==0.2.12
debtcollector==1.22.0
defusedxml==0.6.0
difftacular==0.1
distro==1.4.0
dkimpy==1.0.4
dnspython==1.16.0
# https://code.launchpad.net/~launchpad/dohq-artifactory/+git/dohq-artifactory/+ref/0.7.630
# lp1 Fix the version determination code in setup.py.
dohq-artifactory==0.7.630+lp1
dulwich==0.19.16
eggtestinfo==0.3
enum34==1.1.6
fastimport==0.9.8
feedparser==5.2.1
FormEncode==1.3.1
futures==3.3.0
geoip2==2.9.0
grokcore.component==3.1
gunicorn==19.8.1
httplib2==0.8
hyperlink==18.0.0
immutables==0.14
importlib==1.0.2
importlib-metadata==2.1.1
importlib-resources==0.5
incremental==21.3.0
ipython==0.13.2
iso8601==0.1.12
jmespath==0.10.0
jsautobuild==0.2
keyring==0.6.2
keystoneauth1==4.1.0
kombu==4.6.11
launchpad-buildd==206
launchpadlib==1.10.14
lazr.batchnavigator==1.3.1
lazr.config==2.2.3
lazr.delegates==2.0.4
lazr.enum==1.2.1
lazr.jobrunner==0.17
lazr.lifecycle==1.2.1
lazr.restful==2.0.1
lazr.restfulclient==0.14.4
lazr.sshserver==0.1.13
lazr.uri==1.0.6
lpjsmin==0.6
m2r==0.1.13
Markdown==3.2.2
martian==1.3.post1
maxminddb==1.5.1
meliae==0.5.1
mistune==0.8.3
monotonic==1.5
msgpack==1.0.2
multipart==0.2.4
netaddr==0.7.19
netifaces==0.11.0
oauth==1.0
oauthlib==3.1.0
oops==0.0.14
oops-amqp==0.1.0
oops-datedir-repo==0.0.24
oops-datedir2amqp==0.1.0
oops-timeline==0.0.3
oops-twisted==0.0.8
oops-wsgi==0.0.15
os-service-types==1.7.0
oslo.config==7.0.0
oslo.i18n==3.25.1
oslo.serialization==2.29.3
oslo.utils==3.42.1
paramiko==2.4.1
Paste==3.3.0
PasteDeploy==2.1.0
pathlib2==2.3.2
patiencediff==0.2.2
pgbouncer==0.0.9
# pkginfo 1.7.0 dropped Python 3.5 support, but we need the features of the
# newer version, and both our and the package's test suite show no
# incompatibilities
pkginfo==1.8.2
prettytable==0.7.2
psutil==5.4.2
psycopg2==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pygettextpo==0.2
# lp:~launchpad/pygpgme/pthread
# lp1 Link against libgpgme-pthread for thread-safety.
pygpgme==0.3+lp1
PyHamcrest==1.9.0
pyinotify==0.9.4
PyJWT==1.7.1
pymacaroons==0.13.0
pymemcache==3.5.0
PyNaCl==1.3.0
pyOpenSSL==17.5.0
pystache==0.5.3
python-dateutil==2.8.1
python-debian==0.1.32
python-keystoneclient==3.21.0
# https://github.com/ziima/python-openid/pull/47
# lp1 Fix normalization of non-ASCII query strings on Python 2.
python-openid2==3.2+lp1
python-swiftclient==3.8.1
PyYAML==5.3.1
rabbitfixture==0.5.1
requests-file==1.4.3
requests-toolbelt==0.9.1
responses==0.9.0
rfc3986==1.5.0
s3transfer==0.3.6
secure-cookie==0.1.0
selenium==3.141.0
service-identity==18.1.0
setproctitle==1.1.7
setuptools-git==1.2
setuptools-scm==3.4.3
simplejson==3.8.2
SimpleTAL==5.2
soupmatchers==0.4
soupsieve==1.9
statsd==3.3.0
stevedore==1.32.0
# lp:~launchpad-committers/storm/lp
storm==0.24+lp417
subprocess32==3.2.6
tenacity==6.1.0
testresources==0.2.7
testscenarios==0.4
testtools==2.5.0
timeline==0.0.7
treq==18.6.0
# lp:~launchpad/twisted:lp-backport
Twisted==20.3.0+lp9
txfixtures==0.4.3
txpkgupload==0.4
urllib3==1.25.11
van.testing==3.0.0
vine==1.3.0
virtualenv-tools3==2.0.5
wadllib==1.3.6
waitress==1.3.1
webencodings==0.5.1
WebOb==1.8.5
WebTest==2.0.35
Werkzeug==1.0.1
wheel-filename==1.1.0
wrapt==1.12.1
wsgi-intercept==1.9.2
WSGIProxy2==0.4.6
wsgiref==0.1.2
z3c.pt==3.2.0
z3c.ptcompat==2.2.0
zipp==1.2.0
zope.app.http==4.0.1
zope.app.publication==4.3.1
zope.app.publisher==4.2.0
zope.publisher==6.0.2+lp1
zope.testbrowser==5.5.1
# lp:~launchpad-committers/zope.testrunner:launchpad
zope.testrunner==5.3.0+lp1
zope.vocabularyregistry==1.1.1
zstandard==0.15.2
